+++ packages/http-server-csharp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log - @typespec/http-server-csharp
 
+## 0.58.0-alpha.27
+
+### Bug Fixes
+
+- [#10175](https://github.com/microsoft/typespec/pull/10175) Emit C# class for models that extend another model with no additional properties
+
+
 ## 0.58.0-alpha.26
 
 ### Bump dependencies